feat: グラフをusers.map内に配置するよう修正 #123
軸不要のため
